refactor: extract reusable field test fixtures
